Improve situation with pins
