+++ src/components/map/Map.js
@@ -55,6 +55,7 @@ const Map = ({
   geolocation,
   locations,
   activeLocationId,
+  editingLocationId,
   clusters,
   onGeolocationClick,
   onClusterClick,
@@ -241,6 +242,7 @@ const Map = ({
             lat={location.lat}
             lng={location.lng}
             selected={location.id === activeLocationId}
+            editing={location.id === editingLocationId}
             label={showLabels ? location.typeName : undefined}
           />
         ))}
