+++ src/components/map/MapPage.js
@@ -15,10 +15,10 @@ import { getAllLocations, viewChangeAndFetch } from '../../redux/viewChange'
 import { bootstrapURLKeys } from '../../utils/bootstrapURLKeys'
 import { useAppHistory } from '../../utils/useAppHistory'
 import AddLocationButton from '../ui/AddLocationButton'
-import AddLocationPin from '../ui/AddLocationPin'
 import LoadingIndicator from '../ui/LoadingIndicator'
 import { ConnectedGeolocation } from './ConnectedGeolocation'
 import Map from './Map'
+import { AddLocationPin, EditLocationPin } from './Pins'
 import TrackLocationButton from './TrackLocationButton'
 
 const BottomLeftLoadingIndicator = styled(LoadingIndicator)`
@@ -141,7 +141,7 @@ const MapPage = ({ isDesktop }) => {
       ) : (
         !isDesktop && <AddLocationButton onClick={handleAddLocationClick} />
       )}
-      {isEditingLocation && <AddLocationPin />}
+      {isEditingLocation && <EditLocationPin />}
       {!isDesktop && <TrackLocationButton isIcon />}
 
       {locationRequested && <ConnectedGeolocation />}
@@ -156,6 +156,7 @@ const MapPage = ({ isDesktop }) => {
           typeName: getCommonName(location.type_ids[0]),
         }))}
         activeLocationId={locationId || hoveredLocationId}
+        editingLocationId={isEditingLocation ? locationId : null}
         onViewChange={(newView) => {
           dispatch(viewChangeAndFetch(newView))
         }}
